Split the application up into multiple files
